Renamed wrapper -> binding, LLVM format
As suggested `wrapper` was substituted by `binding`.
The minimum required version and project naming in `CMakeLists.txt`
were removed, as this is not necessary.
