+++ clients/CMakeLists.txt
@@ -5,5 +5,5 @@
 # SPDX-License-Identifier: Apache-2.0
 
 add_subdirectory(opal-rt/rtlab-asyncip)
-add_subdirectory(python-wrapper)
+add_subdirectory(python-binding)
 add_subdirectory(shmem)
