+++ clients/python-binding/CMakeLists.txt
@@ -1,11 +1,8 @@
-# Python-wrapper CMakeLists.
+# Python-binding CMakeLists.
 #
 # Author: Kevin Vu te Laar <vu.te@rwth-aachen.de>
 # SPDX-FileCopyrightText: 2014-2025 Institute for Automation of Complex Power Systems, RWTH Aachen University
 # SPDX-License-Identifier: Apache-2.0
-
-cmake_minimum_required(VERSION 3.15...3.29)
-project(villas-python-wrapper LANGUAGES CXX)
 
 set(PYBIND11_FINDPYTHON ON)
 find_package(pybind11 CONFIG)
@@ -24,12 +21,12 @@ if(pybind11_FOUND)
   message(STATUS "Python minor version: ${Python_VERSION_MINOR}")
   message(STATUS "Python .so install directory: ${PYTHON_LIB_DYNLOAD_DIR}")
 
-  pybind11_add_module(python-wrapper villas-python-wrapper.cpp)
-  set_target_properties(python-wrapper PROPERTIES OUTPUT_NAME villas_node)
-  target_link_libraries(python-wrapper PUBLIC villas)
+  pybind11_add_module(python-binding villas-python-binding.cpp)
+  set_target_properties(python-binding PROPERTIES OUTPUT_NAME villas_node)
+  target_link_libraries(python-binding PUBLIC villas)
 
   install(
-    TARGETS python-wrapper
+    TARGETS python-binding
     COMPONENT lib
     RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
     LIBRARY DESTINATION ${PYTHON_LIB_DYNLOAD_DIR}
